LUCENE-3619: backport assertions from trunk
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_3x@1210471 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/index/NormsWriter.java
@@ -164,7 +164,7 @@ final class NormsWriter extends InvertedDocEndConsumer {
             normsOut.writeByte(defaultNorm);
         }
 
-        assert 4+normCount*state.numDocs == normsOut.getFilePointer() : ".nrm file size mismatch: expected=" + (4+normCount*state.numDocs) + " actual=" + normsOut.getFilePointer();
+        assert 4+normCount*(long)state.numDocs == normsOut.getFilePointer() : ".nrm file size mismatch: expected=" + (4+normCount*(long)state.numDocs) + " actual=" + normsOut.getFilePointer();
       }
       success = true;
     } finally {
